refactor [CS] move actions into package
