+++ app/controllers/AddTeamMemberToGroupsController.scala
@@ -18,6 +18,7 @@ package controllers
 
 import config.AppConfig
 import connectors.AddMembersToAccessGroupRequest
+import controllers.actions.TeamMemberAction
 import forms.AddGroupsToClientForm
 import models.TeamMember
 import play.api.Logging
