+++ app/controllers/actions/TeamMemberAction.scala
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.actions
 
+import controllers._
 import config.AppConfig
 import connectors.AgentPermissionsConnector
 import models.TeamMember
