+++ test/controllers/CreateGroupControllerSpec.scala
@@ -18,6 +18,7 @@ package controllers
 
 import com.google.inject.AbstractModule
 import connectors.{AgentPermissionsConnector, AgentUserClientDetailsConnector}
+import controllers.actions.AuthAction
 import helpers.Css.{H1, H2, paragraphs}
 import helpers.{BaseSpec, Css}
 import models.{AddTeamMembersToGroup, DisplayClient, TeamMember}
